// src/app/interfaces/home/denominacion.ts
export interface Denominacion {
    tituloOtorgar: string;
    registroCalificado: string;
    registroCalificadoUrl: string;
    perAdmision: string;
    numCreditos: string;
    duracion: string;
    costoMatricula: string;
    cupo: string;
    fechaEdicion: string;
    mision: string;
    vision: string;
    proposito: string;
    key$?: string;
}